CDPCAM-48, CDPCAM-43: Custom output and links to docs.
This commits provides the capability to do different output types
(tex and json for now) and provides the different issues encountered
with the ability to have links and different formatting.

It does this through centralizing the Issue messages into
issue_templates.yml files which get loaded at runtime to present the issues
to the renderers.  The renderers pick the format mechanism defined by the
render_type in the issue_templates.yml.

The issue ids can be generated into constants using the
scripts/update_issue_templates.py script.  This will also be run during
the pre-commit.

Signed-off-by: Brian Towles <btowles@cloudera.com>